v4.2.0

+++ htdocs/core/delete.php
@@ -1,9 +1,17 @@
 <?php
 include "config.php";
+session_start();
+
+if (!isset($_SESSION['user_id'])) {
+    echo "This feature is for members only";
+    exit;
+}
 
 if(isset($_GET['id'])){
     $delete_id = mysqli_real_escape_string($conn, $_GET['id']);
+
     $sql = mysqli_query($conn, "DELETE FROM url WHERE shorten_url = '{$delete_id}'");
+
     if($sql){
         echo "success";
     } else {
